requirements.txt
# Requirements for Log Analyzer
websocket-client
requests
llama-index
ollama
llama-index-llms-ollama
llama-index-vector-stores-faiss